+++ src/Rules/ColorRgba.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace scybulski\Validator\Color\Rules;
+
+use Illuminate\Contracts\Validation\Rule;
+
+class ColorRgba implements Rule
+{
+    /**
+     * Determine if the validation rule passes.
+     *
+     * @param  string  $attribute
+     * @param  mixed  $value
+     * @return bool
+     */
+    public function passes($attribute, $value)
+    {
+        preg_match(
+            '/^(rgba)\(([01]?\d\d?|2[0-4]\d|25[0-5])\W+([01]?\d\d?|2[0-4]\d|25[0-5])\W+([01]?\d\d?|2[0-4]\d|25[0-5])\)?\W+([01](\.\d+)?)\)$/i',
+            $value,
+            $m
+        );
+        return count($m) > 0;
+    }
+
+    /**
+     * Get the validation error message.
+     *
+     * @return string
+     */
+    public function message()
+    {
+        return 'The given value is not a valid RGBA color.';
+    }
+}