+++ src/ServiceProvider.php
@@ -1,22 +1,19 @@
 <?php
 
-namespace Tylercd100\Validator\Color;
+namespace scybulski\Validator\Color;
 
 use Illuminate\Support\ServiceProvider as IlluminateServiceProvider;
-use Tylercd100\Validator\Color\Validator as ColorValidator;
+use scybulski\Validator\Color\Rules\Color;
+use scybulski\Validator\Color\Rules\ColorHex;
+use scybulski\Validator\Color\Rules\ColorKeyword;
+use scybulski\Validator\Color\Rules\ColorLongHex;
+use scybulski\Validator\Color\Rules\ColorRgb;
+use scybulski\Validator\Color\Rules\ColorRgba;
+use scybulski\Validator\Color\Rules\ColorRgbAny;
+use scybulski\Validator\Color\Rules\ColorShortHex;
 
 class ServiceProvider extends IlluminateServiceProvider
 {
-    /**
-     * Bootstrap the application events.
-     *
-     * @return void
-     */
-    public function boot()
-    {
-
-    }
-
     /**
      * Register the service provider.
      *
@@ -26,26 +23,36 @@ class ServiceProvider extends IlluminateServiceProvider
     {
         $this->app->resolving('validator', function ($factory, $app) {
 
-            $colorValidator = new ColorValidator();
-
-            $factory->extend('color', function ($attribute, $value, $parameters, $validator) use ($colorValidator) {
-                return $colorValidator->isColor($value);
+            $factory->extend('color', function ($attribute, $value, $parameters, $validator) {
+                return (new Color())->passes($attribute, $value);
             });
 
-            $factory->extend('color_hex', function ($attribute, $value, $parameters, $validator) use ($colorValidator) {
-                return $colorValidator->isColorAsHex($value);
+            $factory->extend('color_hex', function ($attribute, $value, $parameters, $validator) {
+                return (new ColorHex())->passes($attribute, $value);
             });
 
-            $factory->extend('color_rgb', function ($attribute, $value, $parameters, $validator) use ($colorValidator) {
-                return $colorValidator->isColorAsRGB($value);
+            $factory->extend('color_short_hex', function ($attribute, $value, $parameters, $validator) {
+                return (new ColorShortHex())->passes($attribute, $value);
             });
 
-            $factory->extend('color_rgba', function ($attribute, $value, $parameters, $validator) use ($colorValidator) {
-                return $colorValidator->isColorAsRGBA($value);
+            $factory->extend('color_long_hex', function ($attribute, $value, $parameters, $validator) {
+                return (new ColorLongHex())->passes($attribute, $value);
             });
 
-            $factory->extend('color_keyword', function ($attribute, $value, $parameters, $validator) use ($colorValidator) {
-                return $colorValidator->isColorAsKeyword($value);
+            $factory->extend('color_rgb', function ($attribute, $value, $parameters, $validator) {
+                return (new ColorRgb())->passes($attribute, $value);
+            });
+
+            $factory->extend('color_rgba', function ($attribute, $value, $parameters, $validator) {
+                return (new ColorRgba())->passes($attribute, $value);
+            });
+
+            $factory->extend('color_rgb_any', function ($attribute, $value, $parameters, $validator) {
+                return (new ColorRgbAny())->passes($attribute, $value);
+            });
+
+            $factory->extend('color_keyword', function ($attribute, $value, $parameters, $validator) {
+                return (new ColorKeyword())->passes($attribute, $value);
             });
         });
     }